requirements.txt
watchdog
pandas
openpyxl
sqlalchemy
pyodbc
requests
APScheduler
python-dotenv
